Injector API ok and tests for Maps, List, Double and Strings
